Mostly refactoring1
