+++ client-next/src/components/NavBarContainer.tsx
@@ -1,14 +1,13 @@
-import Container from 'react-bootstrap/Container';
 import Navbar from 'react-bootstrap/Navbar';
 export function NavBarContainer() {
   return (
     <Navbar expand="lg" data-bs-theme="light">
-      <Container>
+      <div className="navbar-container">
         <Navbar.Brand>
           <img alt="" src="/img/otp-logo.svg" width="30" height="30" className="d-inline-block align-top" /> OTP Debug
           Client
         </Navbar.Brand>
-      </Container>
+      </div>
     </Navbar>
   );
 }
